:construction: refactor upstream checks endpoint to reduce conflict surface with upstream
