+++ nemoguardrails/server/schemas/openai.py
@@ -171,41 +171,3 @@ class OpenAIModelsList(BaseModel):
     """Standard OpenAI models list response."""
 
     data: list[OpenAIModel] = Field(..., description="List of OpenAI model objects.")
-
-
-class RailStatus(BaseModel):
-    """Status of a single rail execution."""
-
-    status: str = Field(description="Status of the rail: 'success' or 'blocked'.")
-
-
-class MessageCheckResult(BaseModel):
-    """Per-message guardrail check result."""
-
-    index: int = Field(description="Index of the message in the request.")
-    role: str = Field(description="Role of the message (user/assistant/tool).")
-    rails: dict[str, RailStatus] = Field(
-        default_factory=dict,
-        description="Rails that were evaluated for this message and their statuses.",
-    )
-
-
-class GuardrailCheckResponse(BaseModel):
-    """Response body for the /v1/guardrail/checks endpoint."""
-
-    status: str = Field(
-        description="Overall status: 'success' if all rails passed, 'blocked' if any rail blocked, 'error' for system errors."
-    )
-    rails_status: dict[str, RailStatus] = Field(
-        default_factory=dict,
-        description="Status of each individual rail that was executed (aggregated across all messages).",
-    )
-    messages: List[MessageCheckResult] = Field(
-        default_factory=list,
-        description="Per-message guardrail check results showing which rails were evaluated for each message. "
-        "This is a NeMo extension beyond the base OpenAPI spec for enhanced debugging and visibility.",
-    )
-    guardrails_data: Optional[dict] = Field(
-        default=None,
-        description="Additional data from guardrail execution including logs and statistics.",
-    )
